Split track function
