+++ src/symbols-count.js
@@ -4,6 +4,7 @@ import FormData from 'sketch-polyfill-fetch/lib/form-data';
 
 import {SECRET_NAME, TOKEN_NAME, API_ENDPOINT, CLIENT_ID} from './config';
 import {refresh} from './authenticate';
+import track from './track';
 
 const doc = Sketch.getSelectedDocument();
 const pages = doc.pages;
@@ -92,41 +93,3 @@ function shouldCountSymbol(layer) {
   }
   return false;
 }
-
-const track = async ({
-  file_name,
-  file_id,
-  score_percentage,
-  total_symbols_count,
-}) => {
-  const storedSecret = Sketch.Settings.settingForKey(SECRET_NAME);
-  const storedToken = Sketch.Settings.settingForKey(TOKEN_NAME);
-
-  const res = await fetch(
-    `${API_ENDPOINT}/design-platform-analytics/file-score`,
-    {
-      method: 'POST',
-      body: JSON.stringify({
-        file_name: file_name,
-        file_id: file_id,
-        score_percentage: score_percentage,
-        total_symbols_count: total_symbols_count,
-      }),
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${storedToken}`,
-      },
-    }
-  );
-
-  if (!res.ok) {
-    const {_value} = res.json();
-    if (_value.code === 'unauthorized.bad_access_token.expired') {
-      return refresh();
-    }
-  } else {
-    return {
-      result: 200,
-    };
-  }
-};
